Begin implementing lobby manager
Matchmaker should emit an event when a match needs to be created.
Master matchmaker and lobby manager objects can probably be created in app.js. That way event handling can be a bit more centralized. They can probably be moved to a separate file if this ends up making app.js too cluttered.
commands/pvp.js currently needs access to the matchmaker object. If it will be created in app.js it either needs to be passed to an init function for the pvp methods, or it might need to be in a separate file after all.